Added async=true to GuiceFilter in web.xml
See this message in Tomcat log:

WARNING:
=================================================================
Vaadin depends on Atmosphere 2.1.2.vaadin3 but version 2.1.6 was found.
This might cause compatibility problems if push is used.
=================================================================

App loads ok, but takes a full minute or more to respond to a page change - once it gets there, it seems fine, and push works