Fix #311 Upgrade to Vaadin 7.3.9. Also corrects fix to #305
Upgraded to Vaadin 7.3.9.  The earlier fix to #305 was incomplete, but is now correct.  The registeredAnnotations and registeredvalueAnnotations are now bound using annotations independent of I18N and I18NValue.  This removes, for example, the conflict of having an I18N annotation on a class where there also needs to be another @BindingAnnotation