Fix #318. Introduce navigation rules to make Navigator response to login & logout more flexible
Create a Guice NavigationModule.  Move Navigator & URIHandler bindings from UIModule to NavigationModule.  Change UserStatusListener API to provide change of status with login / logout already identified.
The source of login / logout is passed to the rules to enable decisions based on what the source is.  The logic of the login rule is slightly different to the original code, it no longer expects current page to be the login page, and thus supports an embedded login component.