Fix #323 Resolve failure on non-English platform
Class bundle now TestLbls_en added to enable tests to run when CurrentLocale is not Locale.UK as it usually is.  Noted that errors also occur if there is a properties file and a class of the same (fully qualified name).  This has been raised as #325 but may not be worth fixing.  A note has been made in the documentation.

Some logging has been added, and some Javadoc corrctions also made
DefaultFileSitemapLoaderTest modified - it had been using a lable key which now clashes with renamed test resources