Fix #327 Immutable MasterSitemapNode
Rather than make multiple copies of MasterSitemapNodes (one per UserSitemap), MasterSitemapNode is now immutable.  This has required some changes to the process of constructing them, as the logic for that is dispersed.  In the process of testing this, found a bug in the way the Navigator scope is handled (see #342).  This commit therefore fails the Login_Navigation_Rule test in TestApp, pending a fix for #342