Fix #342 Scopes corrected
DefaultNavigator made @UIScoped
DefaultSitemapService made @Singleton

Handling of standard pages in the sitemap improved.  There is no longer a need to explicitly add standard pages, they are detected from the page key.

UserSitemap always holds all standard pages keys, plus an additional mapping of uri to page key.  This is required because not all nodes will be in the UserSitemap, and the uri() method may not return a result.

The is***Uri methods move to Sitemap from UserSitemap