See #348 ServiceListener replaced by event bus
All services (through AbstractService) are set up to publish status changes through the @GlobalBus.  The ServicesMonitorModule is renamed ServiceModule, as it now does more to initialise services - an injection listener intercepts and injects the global event bus, and subscribes the service to the bus.

There is no longer a need to register services with the ServicesMonitor, it is done during service init().