Fix #370 Scoped UI init order
CurrentLocale was being set up from the current environment before the UI had been constructed - this only occurred when the UI deferred component construction to the doLayout method, as opposed to doing it in its own constructor.  The order has been changed in the ScopedUI.init() method so that locale changes do not happen until after the layout has been completed.