See #374 UIModule provides defaults
UIModule was previously abstract and had to be sub-classed for every application.  It has been renamed DefaultUIModule, and uses DefaultApplicationUI to remove the need for sub-classing.  Other potential areas for simplification still to be looked at include the reworking of ScopedUIProvider so that it does not need to be sub-classed for a different UI class;  the application title is currently specified in DefaultUIModule - perhaps this should be moved to the ServletModule, as that has to be defined for each application.