Fix #373 Exclude pages from navigation
The positionIndex for pages is utilised for identifying pages which should not be displayed in navigation components. A positionIndex of < 0 is excluded.  NoNavFilter is used for exclusion during the process of loading the navigation components.

UserSitemapNode now carries a copy of the positionIndex (as opposed to referencing the MasterSitemapNode), to offer the developer the potential to provide a different presentation order in navigation components, depending on some characteristic of the user.

Fix #256 Position index added to Sitemap loaders

The Annotation and Direct sitemap modules and loaders handle a position index for pages which allows the order of pages to be specified (the position index is relative, that is, it is used to sort the page order and it does not therefore matter whether there are gaps in numbering). The File based sitemap has been deprecated and therefore not modified