Fix #384 ID inconsistently applied when class is enhanced
An enhanced class name contains, for example, $$EnhancerByGuice3f556345 when taken from an instance.  This meant that ID.getId() and ID.getIdc()  return different results for an enhanced class.  Fixed by using ClassnameUtil to return the class name without the enhanced portion.
Also, DefaultErrorView logs the error as well as displaying it