See #374 Changes to ScopedUIProvider to reduce new application setup
ScopedUIProvider used to require sub-classing for any application which chose to implement its own UI, and the new sub-class required binding in a sub-class of DefaultUIModule.  There is now a simpler way to bind the UI class to the underlying Vaadin UIProvider with a one line change in a sub-class of  DefaultUIModule.  Sub-classing of ScopedUIProvider is now only required for applications which use multiple UI classes