See #283 ViewBase reconstructs components only when needed
ViewBase delegates to a doBuild for sub-classes to implement, and uses a flag to decide whether a rebuild is necessary or not.  This removes the situation where the field components are reconstructed every time a View is selected.  It is still possible to force a rebuild by clearing the flag in the beforeBuild() method.

At loadData() method has been added, and is called by a @Handler for the AfterViewChangeBusMessage - it is empty by default, and may not be needed when a view is based on Vaadin containers.

The unnecessary translation of keys has not yet been addressed.