See #283.  loadData() was called even when out of focus
Was using the UIBus to trigger a call to loadData, but that means that all views that have been created wihtin a UIScope would be called to load data, even if they are no longer active.  Modified to use direct interface call (ultimately from Navigator) so that load data is only called when the view becomes active.