Fix #396 BasicForest replaceNode() method
The Jung graph that BasicForest is based on ensures that integrity is maintained at all times - the down side to that is it makes it hard to move a node from part of the graph to another, which BasicForest was previously doing incorrectly.  The method now used takes a sub-graph onto the replacement node, and then copies the sub-graph back into the main graph.  This could have some performance issues for a large site, see #397

SitemapChecker has been renamed, as it does actually makes changes to the MasterSitemap, and the name suggested something entirely passive.  It is now SitemapFinisher