See #401 Options Popup available
OptionPopupWindow provides the user with a way to change their options.  Content is generated by reflection from the OptionConext.  OptionKey is generified, and the default value provided as part of the key.  Fix #358 OptionCache fixed - it was not handling writes correctly.  OptionCacheKey modified to always include the userId as part of that fix.

OptionPopup still to be integrated with User Access Control