Fix #401 OptionPopup completed, with user access control integration
The OptionPopup reads an OptionContext for OptionKeys, creates an editor for them dynamically, and allows the user to change the option values. DefaultRealm gives permission for a user to modify their own options only.
The popup includes 'restore to default' buttons.  These delete the user level option entry, thus defaulting to next highest ranked entry.
Removed unnecessary calls with UserHierarchy as a parameter.  These are no longer required, as the hierarchy is injected into the Option implementation.
The DefaultRealm also now has an 'admin' user, which can modify any user options.  DefaultOption checks for permission on write and delete.  There is no restriction on reading options.

Fix #407  An AuthorizationException is now caught and handled via a user notification - although actually the "bug" was caused by throwing the wrong exception in DefaultOption - that has been corrected an now throws an UnauthorizedException.  The DefaultLoginView is modified to throw an exception specifcally for empty username or password, and this has also been added to the DefaultExceptionHandler.

Fix #408  SubjectIdentifier now returns a "?" when no userId is present