Fix #421 Subject and VaadinSession fixed
The MethodInterceptor implementations for all the Shiro annotations have been replaced in order to use SubjectProvider.get() instead of SecurityUtils.getSubject().  The ShiroAopModule is no longer used, and is replaced by the KrailShiroAopModule.

SubjectProvider is an interface now, to overcome circular dependencies in Guice

DefaultSubjectProvider now properly manages the relationship between VaadinSession and Subject.

NotAUserException and NotAGuestException added to more clearly identify cause of authorisation failure, if required.  Handlers for both added, and incorproated into KrailErrorHandler.

Fix #418 StandardShiroModule is now DefaultShiroModule, and a fluent bindRealm() added