See #244.  MasterSitemap thread safety
Introduced SitemapQueue, which is not yet necessary but will help with #254, as it enables the switching of MasterSitemap versions.  A timeout has been added for loading of the MasterSitemap.

MasterSitemap is now locked after initial load, which removes the need for managing concurrent thread access, as it is now read-only once loaded.

A common BusProvider interface has been added

ApplicationConfiguration considered for making Threadsafe.   See #495 for explanation.  Tutorial and Javadoc updated

A number of small changes made as part of quality review