Fix #520 Merged ViewBase18N into ViewBase
Added interface NamedAndDescribed with name and description I18NKeys.  KrailView and Service extend this interface and ViewBase / AbstractService provide the translation.
This means that all views are now constructed with a Translate parameter
ScopedUI makes use of this to add view name to the page title
MockTranslate added to simplify testing with I18N